feat: daemon-based swallowing
This tool now is a daemon. This essentialy works by storing the PID of all X.org windows, and when it is time to swallow, simply save the parent PID - it may vary according to time in some operating systems, so that's the way I found to work. I'm too lazy to explain everything so it's better to read the code yourself.